split middleware tests into individual files
