+++ __tests__/companies-model_update_company.spec.js
