+++ __tests__/middleware_ValidateUserById.spec.js
@@ -1,0 +1,42 @@
+const request = require('supertest');
+const server = require('../api/server');
+const db = require('../data/dbConfig');
+const jwt = require('jsonwebtoken');
+
+describe('server.js', () => {
+  beforeEach(async () => {
+    await db.raw('truncate table users cascade');
+    await db.raw('truncate table reviews cascade');
+    await db.raw('truncate table companies cascade'); //clears table for testing purposes before each test
+  });
+
+  /*************************** VALIDATE USER BY ID *******************************/
+
+  describe('ValidateUserById Middleware', () => {
+    it('should 404 error if the user id does not exist', async () => {
+      // register a new user
+      res = await request(server)
+        .post('/api/auth/register')
+        .send({
+          username: 'mario',
+          password: 'superstar',
+          email: 'mario@gmail.com'
+        });
+      //open the database and see that the new user is there
+      const newUsers = await db('users');
+      expect(newUsers).toHaveLength(1);
+      expect(res.status).toEqual(201);
+      // check token exists
+      const token = res.body.token;
+      expect(token.length).toBeGreaterThan(20);
+      // find a user by id
+      res = await request(server)
+        .get('/api/users/2')
+        .set({ authorization: token, Accept: 'application/json' });
+      expect(res.status).toEqual(404);
+      expect(res.body).toMatchObject({
+        errorMessage: 'The user with the specified ID does not exist'
+      });
+    });
+  });
+});
