+++ __tests__/server.spec.js
@@ -5,7 +5,5 @@ const db = require('../data/dbConfig');
 describe('server.js', () => {
   it('runs the tests', () => {
     expect(true).toBe(true);
-
-    //add more tests here
   });
 });
